83 rename listexpand, extract styles, create stories
