+++ src/storybook/stories/index.js
@@ -63,6 +63,9 @@ export {
 } from '../../Components/IconButton/IconButton.stories.js';
 export { default as List } from '../../Components/List/List.stories.js';
 export {
+  default as ListExpand,
+} from '../../Components/List/ListExpand/ListExpand.stories.js';
+export {
   default as ListItem,
 } from '../../Components/List/ListItem/ListItem.stories.js';
 export {
